moving keys, and switch to autopairs
first i have started to move from blink.pairs back to autopairs.
autopairs just works better and is able to NOT get in my way.
blink.pairs has some issues with being stupid about adding closing
pairs.

next is the movement of keys out from the individual plugins into a
centeral keys config file that is the last thing to load. this is a work
in progress.